refactor(jest): fix coverage patterns for files
