+++ jest.config.js
@@ -1,4 +1,11 @@
 module.exports = {
   setupFilesAfterEnv: ['./jest.setup.js'],
-  coverageDirectory: "coverage",
+  coverageDirectory: 'coverage',
+  collectCoverageFrom: ['src/**/*.js'],
+  coveragePathIgnorePatterns: [
+    'src/index.js',
+    'src/components/layouts/',
+    'src/components/app.js',
+    'src/components/routes.js',
+  ],
 };
