+++ src/components/layouts/Content.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import { Container } from '@material-ui/core';
 import { makeStyles } from "@material-ui/core/styles";
-import Routes from "../Routes";
+import Routes from "../routes";
 
 const useStyles = makeStyles(theme => ({
   appBarSpacer: theme.mixins.toolbar,
